+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fraction::new` now contains a `debug_assert!` ensuring the denominator is
   non-zero.
 
+### Added
+
+- `Rect` now implements `FloatConversion` when its unit type implements it.
+
 ## v0.4.0 (2024-07-22)
 
 ### Breaking Changes